+++ server/src/common/actions/sifa.actions/sifa.actions.ts
@@ -7,7 +7,7 @@ import { SIFA_FIELDS } from "./sifaCsvFields.js";
 import { getCodePostalInfo } from "../../apis/apiTablesCorrespondances.js";
 import { CODES_STATUT_APPRENANT } from "../../constants/dossierApprenantConstants.js";
 import { AuthContext } from "../../model/internal/AuthContext.js";
-import { requireManageEffectifsPermission } from "@/src/http/routes/specific.routes/organisme.routes.js";
+import { requireManageEffectifsPermission } from "../helpers/permissions.js";
 
 export const isEligibleSIFA = ({ historique_statut }) => {
   const filtered = historique_statut.filter(({ date_statut }) => {
