+++ server/src/http/routes/specific.routes/organisme.routes.ts
@@ -8,33 +8,7 @@ import {
 import { findEffectifs } from "../../../common/actions/effectifs.actions.js";
 import { isEligibleSIFA } from "../../../common/actions/sifa.actions/sifa.actions.js";
 import { AuthContext } from "../../../common/model/internal/AuthContext.js";
-import { OrganisationOrganismeFormation } from "../../../common/model/organisations.model.js";
-import { findOFLinkedOrganismesIds } from "../../../common/actions/helpers/permissions.js";
-
-export async function canManageEffectifs(ctx: AuthContext, organismeId: string): Promise<boolean> {
-  const organisation = ctx.organisation;
-  switch (organisation.type) {
-    case "ORGANISME_FORMATION_FORMATEUR":
-    case "ORGANISME_FORMATION_RESPONSABLE":
-    case "ORGANISME_FORMATION_RESPONSABLE_FORMATEUR": {
-      const linkedOrganismesIds = await findOFLinkedOrganismesIds(ctx as AuthContext<OrganisationOrganismeFormation>);
-      return linkedOrganismesIds.map((id) => id.toString()).includes(organismeId);
-    }
-
-    case "OPERATEUR_PUBLIC_NATIONAL":
-    case "ADMINISTRATEUR":
-      return true;
-
-    default:
-      return false;
-  }
-}
-
-export async function requireManageEffectifsPermission(ctx: AuthContext, organismeId: string): Promise<void> {
-  if (!(await canManageEffectifs(ctx, organismeId))) {
-    throw Boom.forbidden("Permissions invalides");
-  }
-}
+import { requireManageEffectifsPermission } from "../../../common/actions/helpers/permissions.js";
 
 export async function getOrganismeEffectifs(ctx: AuthContext, organismeId: string, sifa = false) {
   await requireManageEffectifsPermission(ctx, organismeId);
